Проверка ДЗ | 3. Переменные и объекты
- Правильно: №0, №1, №2, №3, №4, №5, №6, №7, №8, №9, №10
- Неправильно:
- Улучшены: №7, №8,
- На доработку:

+++ tasks/task_8.py
@@ -1,4 +1,4 @@
-celsius: int = int(input("Введите температуру в градусах Цельсия: "))
+celsius: float = float(input("Введите температуру в градусах Цельсия: "))
 
 fahrenheit: float = (celsius * 1.8) + 32
